+++ pipelex/cogt/exceptions.py
@@ -1,4 +1,4 @@
-from pipelex.tools.exceptions import FatalError, RootException
+from pipelex.system.exceptions import FatalError, RootException
 from pipelex.types import StrEnum
