+++ pipelex/cogt/models/model_deck.py
@@ -23,7 +23,7 @@ from pipelex.cogt.llm.llm_setting import (
 from pipelex.cogt.model_backends.model_constraints import ModelConstraints
 from pipelex.cogt.model_backends.model_spec import InferenceModelSpec
 from pipelex.system.configuration.config_model import ConfigModel
-from pipelex.tools.exceptions import ConfigValidationError
+from pipelex.system.exceptions import ConfigValidationError
 from pipelex.types import Self
 
 LLM_PRESET_DISABLED = "disabled"
