+++ pipelex/plugins/openai/vertexai_factory.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from pipelex.cogt.exceptions import CogtError, MissingDependencyError
 from pipelex.system.configuration.config_model import ConfigModel
-from pipelex.tools.exceptions import CredentialsError
+from pipelex.system.exceptions import CredentialsError
 from pipelex.tools.misc.json_utils import load_json_dict_from_path
 from pipelex.tools.misc.placeholder import value_is_placeholder
 from pipelex.types import StrEnum
