+++ pipelex/system/configuration/config_root.py
@@ -3,7 +3,7 @@ from typing import Any
 from pydantic import ValidationError
 
 from pipelex.system.configuration.config_model import ConfigModel
-from pipelex.tools.exceptions import ConfigValidationError
+from pipelex.system.exceptions import ConfigValidationError
 from pipelex.tools.typing.pydantic_utils import format_pydantic_validation_error
 from pipelex.types import StrEnum
