+++ pipelex/system/registries/func_registry.py
@@ -5,7 +5,7 @@ from typing import Any, TypeVar, cast, get_type_hints
 
 from pydantic import Field, PrivateAttr, RootModel
 
-from pipelex.tools.exceptions import ToolException
+from pipelex.system.exceptions import ToolException
 from pipelex.urls import URLs
 
 FUNC_REGISTRY_LOGGER_CHANNEL_NAME = "func_registry"
