+++ pipelex/tools/jinja2/jinja2_errors.py
@@ -1,4 +1,4 @@
-from pipelex.tools.exceptions import ToolException
+from pipelex.system.exceptions import ToolException
 
 
 class Jinja2TemplateSyntaxError(ToolException):
