+++ pipelex/tools/secrets/secrets_utils.py
@@ -2,7 +2,7 @@ import re
 
 from pipelex.hub import get_secrets_provider
 from pipelex.system.environment import EnvVarNotFoundError, get_optional_env, get_required_env
-from pipelex.tools.exceptions import ToolException
+from pipelex.system.exceptions import ToolException
 from pipelex.tools.secrets.secrets_errors import SecretNotFoundError
 from pipelex.types import StrEnum
